Move thread pool impl to header
so it can be shared with HVX
